Create rna_analysis.py
This contains class definitions for optimizing RNA sequences based on the byo-doped-results data from Irene Chen's research group at UCLA (contact me directly if you would like that data). It uses a consensus-based method of SVD classification of whether or not point mutations along the sequence enhance or detriment ribozyme activity, and then combines those classifications to predict sequences that will have highest activity based on the data. It's primary purpose is for independent research of ribozyme data, and it is outfitted with several features for studying how manipulations of the database will affect outcomes.